serviceachat/requirements.txt
asgiref==3.5.2
Django==4.1.2
djangorestframework==3.14.0
python-environ==0.4.54
pytz==2022.5
sqlparse==0.4.3
